pb_dse/requirements2.txt
scikit-allel==1.3.3
lazypredict==0.2.5
pytest-docker==0.10.1
django==2.2.4
celery==4.2.0
django-celery==3.2.2
django-celery-results==1.0.1
djangorestframework
vine==1.3.0
kombu==4.2.0
redis
